added ability to set min amount for OFT and NativeOFT

+++ contracts/mocks/WETH9.sol
@@ -1,24 +1,23 @@
-
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
 contract WETH9 {
     string public name = "Wrapped Ether";
     string public symbol = "WETH";
-    uint8  public decimals = 18;
+    uint8 public decimals = 18;
 
-    mapping (address => uint) public balanceOf;
-    mapping (address => mapping (address => uint))  public allowance;
+    mapping(address => uint) public balanceOf;
+    mapping(address => mapping(address => uint)) public allowance;
 
- 	event Transfer(address indexed _from, address indexed _to, uint256 _value);
-    event Approval(address indexed _owner, address indexed _spender, uint256 _value);
-	event Deposit(address indexed from, uint256 amount);
-    event Withdrawal(address indexed to, uint256 amount);
+    event Transfer(address indexed _from, address indexed _to, uint _value);
+    event Approval(address indexed _owner, address indexed _spender, uint _value);
+    event Deposit(address indexed from, uint amount);
+    event Withdrawal(address indexed to, uint amount);
 
     receive() external payable {
         deposit();
     }
-	
+
     function deposit() public payable {
         balanceOf[msg.sender] += msg.value;
         emit Deposit(msg.sender, msg.value);
@@ -48,7 +47,7 @@ contract WETH9 {
     function transferFrom(address src, address dst, uint wad) public returns (bool) {
         require(balanceOf[src] >= wad, "transferFrom: not enough balance");
 
-        if (src != msg.sender && allowance[src][msg.sender] != type(uint256).max) {
+        if (src != msg.sender && allowance[src][msg.sender] != type(uint).max) {
             require(allowance[src][msg.sender] >= wad, "transferFrom: not enough allowance");
             allowance[src][msg.sender] -= wad;
         }
